fix fee tier not being bigint type

+++ sdk/src/wasm/storage/fee_tier.rs
@@ -33,6 +33,6 @@ impl FeeTier {
 #[wasm_bindgen(js_name = "_newFeeTier")]
 pub fn new_fee_tier(js_fee: JsValue, js_tick_spacing: JsValue) -> Result<JsValue, JsValue> {
     let fee: Percentage = convert!(js_fee)?;
-    let tick_spacing: u16 = convert!(js_tick_spacing)?;
-    resolve!(FeeTier::new(fee, tick_spacing))
+    let tick_spacing: u64 = convert!(js_tick_spacing)?;
+    resolve!(FeeTier::new(fee, tick_spacing as u16))
 }
